feature/34344-update-the-datetime-parser (!301)
the issue was that the filtration from/to date was using the UTC date and the events timeline kept the local time, so the event visible on this ss displayed 8th December 00:59 when it should display 7th December 23:59

I updated that one. Now, we have a choice whether to keep local time (local time is useful for the notification).
Also, I removed methods used to format date/time since we use moment library and it already has those methods implemented in a very nice way

![image](/attachments/3982fdd5-c947-4847-928a-b0842f8ab0b6)

Reviewed-on: https://scm.openanalytics.eu/git/rdepot-client/pulls/301
Reviewed-by: Weronika Mrozek <weronika.mrozek@openanalytics.eu>
Co-authored-by: kkulesa <katarzyna.kulesa@openanalytics.eu>
Co-committed-by: kkulesa <katarzyna.kulesa@openanalytics.eu>