feature/34980-Improve-wording-current-on-package-details-page (!317)
I removed the word "current" completely from the package details. Now, the chart has the "current" version as a subtitle (to avoid confusion with the data presented in the chart). In the package versions, the "current" one is left in bold, as for the information about the latest, we are still waiting for the backend.

Reviewed-on: https://scm.openanalytics.eu/git/rdepot-client/pulls/317
Reviewed-by: Katarzyna Kulesa <katarzyna.kulesa@openanalytics.eu>
Co-authored-by: wmrozek <weronika.mrozek@openanalytics.eu>
Co-committed-by: wmrozek <weronika.mrozek@openanalytics.eu>