feature/34681-Improve-form-validation (!344)
finally! What are the changes? All the forms fully rely on Zod schema validators, which makes it easier to maintain and more readable (since all the validation rules for the given form should be in one schema)
I think there is still a place for the improvements, but having this it will be easier to add them during all other tasks since it shouldn't be huge changes (like more error messages, which are now also easy to add, let's just test where we need more of them)

The biggest change is the upload submission flow - before we had to use separate arrays for each property and then search those arrays in order to find a proper value/error. Now each submission is kept as a separate object, so it's known from the beginning which value is assigned to which file (also field store removed, in this approach we simply don't need it)

Another big change is repository schema, because we need to use repository store in the schema definition (to check if there exists any other repository with the given name) but schemas are created before mounted, so they don't have access to the store. That is why the repository's schema has a slightly different structure. We can use it in the future in any other schema if we need to store the fields definition

Reviewed-on: https://scm.openanalytics.eu/git/rdepot-client/pulls/344
Reviewed-by: Jonas Van Malder <jonas.vanmalder@openanalytics.eu>
Co-authored-by: kkulesa <katarzyna.kulesa@openanalytics.eu>
Co-committed-by: kkulesa <katarzyna.kulesa@openanalytics.eu>